refactor: Receiver 클래스들에서 Validator 클래스를 분리
