+++ src/main/java/lotto/domain/LottoNumber.java
@@ -3,10 +3,8 @@ package lotto.domain;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import lotto.exception.LottoNumberException;
 
 public enum LottoNumber {
 
@@ -22,9 +20,6 @@ public enum LottoNumber {
 
     private static final int PICKUP_COUNT = 6;
     private static final String LOTTO_NUMBER_PREFIX = "NUMBER_";
-    private static final String REGEX_FOR_NATURAL_NUMBER = "^[1-9][0-9]*$";
-    private static final int MINIMUM_RANGE = 1;
-    private static final int MAXIMUM_RANGE = 45;
 
     private final int number;
 
@@ -49,30 +44,8 @@ public enum LottoNumber {
         return Arrays.stream(LottoNumber.values()).collect(Collectors.toList());
     }
 
-    public static LottoNumber getByString(String input) {
-        checkNaturalNumber(input);
-        checkRange(input);
-        return LottoNumber.valueOf(LOTTO_NUMBER_PREFIX + input);
-    }
-
-    private static void checkNaturalNumber(String input) {
-        if (!isNaturalNumber(input)) {
-            throw new LottoNumberException(LottoNumberException.LOTTO_NUMBER_NATURAL_NUMBER_ERROR_MESSAGE);
-        }
-    }
-
-    private static boolean isNaturalNumber(String input) {
-        return Pattern.compile(REGEX_FOR_NATURAL_NUMBER).matcher(input).find();
-    }
-
-    private static void checkRange(String input) {
-        if (!isCorrectRange(Integer.parseInt(input))) {
-            throw new LottoNumberException(LottoNumberException.LOTTO_NUMBER_RANGE_ERROR_MESSAGE);
-        }
-    }
-
-    private static boolean isCorrectRange(int input) {
-        return input >= MINIMUM_RANGE && input <= MAXIMUM_RANGE;
+    public static LottoNumber findByNumber(int number) {
+        return LottoNumber.valueOf(LOTTO_NUMBER_PREFIX + number);
     }
 
     @Override
