+++ src/main/java/lotto/receiver/BonusNumberReceiver.java
@@ -2,19 +2,15 @@ package lotto.receiver;
 
 import lotto.domain.Lotto;
 import lotto.domain.LottoNumber;
-import lotto.exception.BonusNumberException;
+import lotto.receiver.validator.BonusNumberValidator;
+import lotto.receiver.validator.LottoNumberValidator;
 
 public class BonusNumberReceiver {
 
     public static LottoNumber receive(String input, Lotto winningNumbers) {
-        LottoNumber lottoNumber = LottoNumber.getByString(input);
-        checkDuplication(winningNumbers, lottoNumber);
-        return lottoNumber;
-    }
-
-    private static void checkDuplication(Lotto winningNumbers, LottoNumber lottoNumber) {
-        if (winningNumbers.getLotto().contains(lottoNumber)) {
-            throw new BonusNumberException(BonusNumberException.BONUS_NUMBER_DUPLICATION_ERROR_MESSAGE);
-        }
+        LottoNumberValidator.validate(input);
+        LottoNumber bonusNumber = LottoNumber.findByNumber(Integer.parseInt(input));
+        BonusNumberValidator.validate(LottoNumber.findByNumber(Integer.parseInt(input)), winningNumbers);
+        return bonusNumber;
     }
 }
